Refactoring the aplication for implement the main menu2
