--- personal/Condicional.java
@@ -1,16 +1,0 @@
-package com.toadsdewin.CocheProyecto.personal;
-
-public class Condicional
-{
-
-	public static void main (String args[])
-	{
-		boolean A =  false;
-		boolean B =  false;
-
-		System.out.println("A and B: "+(A&&B));
-		System.out.println("A or B: "+(A||B));
-		System.out.println("Not A and Not B: " +(!A && !B));
-		System.out.println("Not A: "+(!A));
-	}
-}
